Moved functions into _body.cpp
