+++ engine/engine/src/engine.cpp
@@ -1305,6 +1305,7 @@ namespace dmEngine
         script_lib_context.m_HidContext      = engine->m_HidContext;
         script_lib_context.m_GraphicsContext = engine->m_GraphicsContext;
         script_lib_context.m_JobThread       = engine->m_JobThreadContext;
+        script_lib_context.m_ConfigFile      = engine->m_Config;
 
         if (engine->m_SharedScriptContext) {
             script_lib_context.m_ScriptContext = engine->m_SharedScriptContext;
