+++ engine/gamesys/src/gamesys/gamesys.h
@@ -156,6 +156,7 @@ namespace dmGameSystem
         dmGraphics::HContext    m_GraphicsContext;
         dmJobThread::HContext   m_JobThread;
         dmScript::HContext      m_ScriptContext;
+        dmConfigFile::HConfig   m_ConfigFile;
     };
